+++ index.html
@@ -4,7 +4,7 @@
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="./style.css">
+    <link rel="stylesheet" href="./index.css">
     <title>Awosome Books</title>
     <link rel="preconnect" href="https://fonts.googleapis.com">
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin>
@@ -67,6 +67,8 @@
           <p>Copyright ...</p>
         </div>
     </footer>
-<script src="./js/index.js"></script>
+
+<script src="https://moment.github.io/luxon/global/luxon.min.js"></script>
+<script type="module" src="./js/index.js"></script>
 </body>
 </html>
